# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qubah Learning API",
        "description": "REST API for the Qubah Learning educational platform. Provides endpoints for authentication, subjects, topics, content, progress tracking, and parent dashboard.",
        "contact": {
            "name": "Qubah Team",
            "email": "support@qubah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/subscriptions/plans": {
            "get": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Get all subscription plans",
                "description": "Display a listing of plans.",
                "operationId": "getPlans",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Create a new subscription plan",
                "description": "Store a newly created plan.",
                "operationId": "createPlan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "duration_months"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Premium Plan"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "duration_months": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "features": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/plans/{id}": {
            "put": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Update a subscription plan",
                "description": "Update the specified plan.",
                "operationId": "updatePlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "duration_months": {
                                        "type": "integer"
                                    },
                                    "features": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Delete a subscription plan",
                "description": "Remove the specified plan.",
                "operationId": "deletePlan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions": {
            "get": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Get all user subscriptions",
                "description": "Display a listing of subscriptions.",
                "operationId": "getSubscriptions",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Assign a subscription package to a user",
                "description": "Assign a subscription to a user.",
                "operationId": "assignSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "user_name",
                                    "plan_name",
                                    "start_date",
                                    "end_date",
                                    "amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "example": "Timmy Student"
                                    },
                                    "plan_name": {
                                        "type": "string",
                                        "example": "Yearly Plan"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 450
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{id}": {
            "put": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Update a user subscription record",
                "description": "Update the specified subscription.",
                "operationId": "updateSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plan_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "expired",
                                            "cancelled"
                                        ]
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Delete a subscription record",
                "description": "Delete the subscription.",
                "operationId": "deleteSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{id}/cancel": {
            "post": {
                "tags": [
                    "Plans & Subscriptions"
                ],
                "summary": "Cancel an active user subscription",
                "description": "Cancel the subscription.",
                "operationId": "cancelSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login via /api/v1/auth/login and use the returned access_token here.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}